Extract isFaviconUrl
