+++ src/lib/getFaviconLinks.ts
@@ -1,24 +1,12 @@
 import { CheerioAPI } from 'cheerio';
 import { buildAbsoluteURL } from 'url-toolkit';
-import isUrl from 'validator/lib/isURL';
 
 import { FaviconLink } from '../types';
 
 import getBaseHref from './getBaseHref';
+import isFaviconUrl from './isFaviconUrl';
 import parseLinkSizes from './parseLinkSizes';
 import sortFaviconLinks from './sortFaviconLinks';
-
-const isFaviconUrl = (url: unknown): boolean => {
-  if (typeof url !== 'string') {
-    return false;
-  }
-
-  if (url.startsWith('/')) {
-    return isUrl('http://example.com' + url, { protocols: ['http', 'https'] });
-  }
-
-  return isUrl(url, { protocols: ['http', 'https'] });
-};
 
 const getFaviconLinks = ($: CheerioAPI, documentHref: string): FaviconLink[] => {
   const $links = [
@@ -35,6 +23,7 @@ const getFaviconLinks = ($: CheerioAPI, documentHref: string): FaviconLink[] => 
       const linkSizes = $($link).attr('sizes');
       const sizes = parseLinkSizes(linkSizes);
       const url = buildAbsoluteURL(baseHref, linkHref);
+
       return { sizes, source: 'html', url };
     });
